requirements.txt
Django==5.0.6
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
psycopg2==2.9.9
django-filter==24.2